simulators/eth2/common: extract builder,clients
